feat: add UniFFI bridge module with SpvClient facade
Add bridge module behind `uniffi` feature flag exposing:
- SpvClient UniFFI object with lifecycle methods (start, stop, shutdown)
- State queries (is_running, tip_height, tip_hash, peer_count, sync_progress)
- SyncEvent and NetworkEvent enums with all variants
- SpvEventListener callback trait for event dispatch
- Custom type mappings for Network, SocketAddr, PathBuf, ClientConfig

+++ dash-spv/src/client/lifecycle.rs
@@ -129,7 +129,7 @@ impl<W: WalletInterface, N: NetworkManager, S: StorageManager> DashSpvClient<W, 
                 wallet.clone(),
                 mempool_state.clone(),
                 config.mempool_strategy,
-                config.max_mempool_transactions,
+                config.max_mempool_transactions as usize,
             ));
         }
